Homework 3/app1/requirements.txt
functions-framework==3.3.0
flask==2.2.5
google-cloud-error-reporting==1.9.1
google-cloud-storage==2.10.0
google-cloud-pubsub==2.18.4
google-cloud-logging==3.8.0